pyproject.toml
[tool.poetry]
name = "moscode-data-collector"
version = "0.1.0"
description = ""
authors = ["smbrine <smbrine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"
uvicorn = "^0.27.0.post1"
alembic = "^1.13.1"
black = "^24.1.1"
pylint = "^3.0.3"
flake8 = "^7.0.0"
prometheus-fastapi-instrumentator = "^6.1.0"
greenlet = "^3.0.3"
asyncpg = "^0.29.0"
redis = "^5.0.1"
aiokafka = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
